@@ -3,15 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@urql/exchange-graphcache": "^2.4.2",
+    "@hot-loader/react-dom": "^17.0.1",
+    "@urql/core": "^2.3.1",
+    "@urql/exchange-auth": "^0.1.6",
+    "@urql/exchange-graphcache": "^4.3.0",
     "ahooks": "^2.10.9",
     "antd": "^3.26.20",
     "classnames": "^2.2.6",
     "core-js": "3",
+    "dayjs": "^1.10.6",
     "dequal": "^1.0.0",
     "graphql": "^14.2.1",
     "i18next": "^15.1.0",
-    "js-cookie": "^2.2.1",
     "jwt-decode": "^2.2.0",
     "moment": "^2.24.0",
     "nanoid": "^2.0.1",
@@ -28,6 +31,8 @@
     "react-monaco-editor": "^0.25.1",
     "react-resizable": "^1.10.1",
     "react-resize-detector": "^4.2.1",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
     "react-sizes": "^2.0.0",
     "react-table": "^7.0.0-beta.12",
     "react-vega": "^7.4.3",
@@ -36,10 +41,11 @@
     "recoil": "^0.4.0",
     "subscriptions-transport-ws": "^0.9.16",
     "unchanged": "^2.1.0",
-    "urql": "^1.9.7",
+    "urql": "^2.0.4",
     "vega": "^5.20.2",
     "vega-lite": "^5.1.0",
     "vega-lite-api": "^5.0.0",
+    "wonka": "^4.0.15",
     "wouter": "^2.7.4"
   },
   "scripts": {
@@ -70,7 +76,7 @@
     "@urql/devtools": "^0.1.2",
     "babel-plugin-module-resolver": "^3.2.0",
     "babel-plugin-prismjs": "^1.1.1",
-    "eslint": "^6.7.2",
+    "eslint": "^7.5",
     "eslint-config-airbnb": "^18.0.1",
     "eslint-config-prettier": "^4.3.0",
     "eslint-plugin-import": "^2.14.0",
