@@ -1,11 +1,14 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 
-import { Row, Col, message } from 'antd';
+import { useUpdateEffect } from 'ahooks';
+import { useTranslation } from 'react-i18next';
+import { Row, Col } from 'antd';
 
-import useDataSources from 'hooks/useDataSources';
+import useSources from 'hooks/useSources';
 import useTableState from 'hooks/useTableState';
-import useDataSourcesSubscription from 'hooks/useDataSourcesSubscription';
+import useCheckResponse from 'hooks/useCheckResponse';
+import useCurrentUserState from 'hooks/useCurrentUserState';
 
 import DataSourceModal from 'components/DataSourceModal';
 import TableList from 'components/TableList';
@@ -13,12 +16,14 @@ import TableList from 'components/TableList';
 import formatDistanceToNow from '../utils/formatDistanceToNow';
 
 const DataSourcesTable = ({ editId, onModalClose, onModalOpen }) => {
+  const { t } = useTranslation();
   const initModal = id => ({
     editId: id,
     visibleModal: !!id,
   });
 
   const [state, setState] = useState(initModal(editId));
+  const { currentUserState: currentUser } = useCurrentUserState();
 
   useEffect(
     () => setState(initModal(editId)),
@@ -45,20 +50,33 @@ const DataSourcesTable = ({ editId, onModalClose, onModalOpen }) => {
       currentData: {
         fetching: currentLoading
       },
-      executeQueryAll,
+      execQueryAll,
+      execQueryCurrent,
     },
     mutations: {
-      editMutation, mExecuteEditMutation,
-    }
-  } = useDataSources({ editId: state.editId, paginationVars });
-
-  useDataSourcesSubscription(() => {
-    executeQueryAll({ requestPolicy: 'network-only' });
+      updateMutation, execUpdateMutation,
+    },
+  } = useSources({
+    params: {
+      editId: state.editId,
+    },
+    pagination: paginationVars,
+    pauseQueryAll: false,
   });
+
+  const userSourcesCount = currentUser.datasources?.length || 0; 
+  const sourcesCount = dataSources?.length || 0; 
+
+  useUpdateEffect(() => {
+    if (sourcesCount && userSourcesCount && userSourcesCount !== sourcesCount) {
+      execQueryAll({ requestPolicy: 'network-only' });
+    }
+  }, [userSourcesCount, sourcesCount, execQueryAll]);
 
   const onDataSourceOpen = (record) => {
     onModalOpen(record);
-    setState(prev => ({ ...prev, editId: record.rowId, visibleModal: true }));
+    setState(prev => ({ ...prev, editId: record.id, visibleModal: true }));
+    execQueryCurrent();
   };
 
   const onDataSourceClose = () => {
@@ -67,27 +85,19 @@ const DataSourcesTable = ({ editId, onModalClose, onModalOpen }) => {
   };
 
   const onSave = (_record, values) => {
-    mExecuteEditMutation(values);
+    execUpdateMutation({ 
+      pk_columns: { id: state.editId },
+      _set: values,
+    });
   };
 
   const onDelete = () => {
     onDataSourceClose();
-    // https://formidable.com/open-source/urql/docs#refetching-data
-    executeQueryAll({ requestPolicy: 'cache-and-network' });
   };
 
-  useEffect(
-    () => {
-      if (editMutation.data) {
-        message.success('Saved');
-      }
-
-      if (editMutation.error) {
-        message.error('Something went wrong');
-      }
-    },
-    [editMutation]
-  );
+  useCheckResponse(updateMutation, () => {}, {
+    successMessage: t('Saved')
+  });
 
   const columns = [
     {
@@ -97,24 +107,24 @@ const DataSourcesTable = ({ editId, onModalClose, onModalOpen }) => {
     },
     {
       title: 'Type',
-      dataIndex: 'dbType',
-      key: 'dbType',
+      dataIndex: 'db_type',
+      key: 'db_type',
     },
     {
       title: 'Updated At',
-      dataIndex: 'updatedAt',
-      key: 'updatedAt',
+      dataIndex: 'updated_at',
+      key: 'updated_at',
       render: (_, record) => {
-        const updatedAt = formatDistanceToNow(record.updatedAt);
+        const updatedAt = formatDistanceToNow(record.updated_at);
         return updatedAt;
       },
     },
     {
       title: 'Created At',
-      dataIndex: 'createdAt',
-      key: 'createdAt',
+      dataIndex: 'created_at',
+      key: 'created_at',
       render: (_, record) => {
-        const createdAt = formatDistanceToNow(record.createdAt);
+        const createdAt = formatDistanceToNow(record.created_at);
         return createdAt;
       },
     },
@@ -127,18 +137,18 @@ const DataSourcesTable = ({ editId, onModalClose, onModalOpen }) => {
       dataSource={dataSource}
       onCancel={onDataSourceClose}
       visible={state.visibleModal}
-      loading={currentLoading || editMutation.fetching}
+      loading={currentLoading || updateMutation.fetching}
       onSave={onSave}
       onDelete={onDelete}
       initialValues={{
-        rowId: editId,
+        id: editId,
       }}
     />,
     <Row type="flex" justify="space-around" align="top" gutter={24} key="1">
       <Col span={24}>
         <TableList
           loading={allLoading}
-          rowKey={row => row.rowId}
+          rowKey={row => row.id}
           columns={columns}
           dataSource={Object.values(dataSources)}
           onRow={record => ({
