@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 
 import { useTranslation } from 'react-i18next';
@@ -10,38 +10,43 @@ import cx from 'classnames';
 import logoImage from 'assets/images/logo.svg';
 import logoBlackImage from 'assets/images/logo-black.svg';
 
-import useAuth from 'hooks/useAuth';
 import useAuthToken from 'hooks/useAuthToken';
-import useDataSourcesSubscription from 'hooks/useDataSourcesSubscription';
-import useDashboardsSubscription from 'hooks/useDashboardsSubscription';
+import useLocation from 'hooks/useLocation';
+import useAppSettings from 'hooks/useAppSettings';
+import useCurrentUserState from 'hooks/useCurrentUserState';
+import useCurrentTeamState from 'hooks/useCurrentTeamState';
+
 import useSider from 'hooks/useSider';
 import usePermissions from 'hooks/usePermissions';
+import useCurrentUser from 'hooks/useCurrentUser';
+
 import MenuView from 'components/MenuView';
 
 import s from './Layout.module.css';
 
 const { Header, Content, Footer, Sider } = Layout;
 
-const UserMenu = ({ mode, restrictScopes = [] }) => {
-  const { logout } = useAuth({});
+const UserMenu = ({ mode, anyTeam, restrictScopes = [] }) => {
+  const { doLogout } = useAuthToken();
   const { t } = useTranslation();
+  const { withAuthPrefix } = useAppSettings();
 
   const accountSubMenu = [
     {
-      path: '/d/profile',
+      path: withAuthPrefix('/profile'),
       title: t('Profile'),
     }
   ];
 
   if (!restrictScopes.includes('team')) {
     accountSubMenu.push({
-      path: '/d/team',
+      path: anyTeam ? withAuthPrefix('/team') : withAuthPrefix('/team/new'),
       title: t('Team'),
     });
   }
 
   accountSubMenu.push({
-    onClick: logout,
+    onClick: doLogout,
     title: t('Logout'),
   });
 
@@ -62,61 +67,95 @@ const UserMenu = ({ mode, restrictScopes = [] }) => {
   );
 };
 
+const TeamMenu = ({ mode, anyTeam, members }) => {
+  const { t } = useTranslation();
+  const { withAuthPrefix } = useAppSettings();
+  const { currentTeamState, setCurrentTeamState } = useCurrentTeamState();
+  const [location] = useLocation();
+
+  if (!anyTeam) {
+    return null;
+  }
+
+  const teamSubMenu = [
+    {
+      path: withAuthPrefix('/team/new'),
+      title: t('Create new team'),
+    }
+  ];
+
+  members?.forEach((member) => {
+    const { team } = member || {};
+    const { name } = team;
+
+    teamSubMenu.unshift({
+      onClick: () => setCurrentTeamState(team),
+      path: location,
+      title: name,
+    });
+  });
+
+  const routes = [
+    {
+      title: currentTeamState?.name,
+      children: teamSubMenu,
+    }
+  ];
+
+  return (
+    <MenuView
+      mode={mode}
+      className={s.header}
+      style={{ lineHeight: '64px' }}
+      nodes={routes}
+    />
+  );
+};
+
 const MainMenu = (props) => {
   const { mode, restrictScopes } = props;
   const { t } = useTranslation();
 
   const {
-    lastUsedDataSourceId,
-    queries: {
-      execCurrentUserQuery,
-    },
-  } = useAuth();
+    currentUserState: currentUser,
+  } = useCurrentUserState();
 
-  const {
-    userState,
-  } = useAuthToken();
+  const { withAuthPrefix } = useAppSettings();
 
-  const dataSources = userState?.allDatasources?.nodes || [];
-  const dashboards = userState?.allDashboards?.nodes || [];
+  const dataSources = currentUser?.datasources || [];
+  const dashboards = currentUser?.dashboards || [];
   const dashboardsCount = dashboards?.length || 0;
 
-  useDataSourcesSubscription(() => {
-    execCurrentUserQuery({ requestPolicy: 'network-only' });
-  });
-
-  useDashboardsSubscription(() => {
-    execCurrentUserQuery({ requestPolicy: 'network-only' });
-  });
-
   const dataSourceItems = dataSources.map(source => ({
-    key: source.rowId,
-    path: `/d/schemas/${source.rowId}`,
+    key: source.id,
+    path: withAuthPrefix(`/schemas/${source.id}`),
     title: source.name,
   }));
 
   const dashboardItems = dashboards.map(dashboard => ({
-    key: dashboard.rowId,
-    path: `/d/dashboards/${dashboard.rowId}`,
+    key: dashboard.id,
+    path: withAuthPrefix(`/dashboards/${dashboard.id}`),
     title: dashboard.name,
   }));
 
-  const lastDataSource = dataSources.find(source => source.rowId == lastUsedDataSourceId);
-  const dataSourceId = lastDataSource?.rowId || dataSources?.[0]?.rowId || '';
+  const lastUsedDataSourceId = null;
+
+  const lastDataSource = dataSources.find(source => source.id == lastUsedDataSourceId);
+  const dataSourceId = lastDataSource?.id || dataSources?.[0]?.id || '';
 
   let routes = [
     {
-      path: `/d/explore/${dataSourceId}`,
+      path: withAuthPrefix(`/explore/${dataSourceId}`),
       title: t('Explore'),
       scope: 'explore/header'
     },
     {
-      path: '/d/sources',
+      path: withAuthPrefix('/sources'),
       title: t('Data Sources'),
       scope: 'datasources'
     },
     {
-      path: '/d/schemas',
+      path: withAuthPrefix('/schemas'),
       title: t('Data Schemas'),
       children: dataSourceItems,
       scope: 'dataschemas'
@@ -125,7 +164,7 @@ const MainMenu = (props) => {
 
   if (dashboardsCount) {
     routes.unshift({
-      path: `/d/dashboards`,
+      path: withAuthPrefix('/dashboards'),
       title: t('Dashboards'),
       children: dashboardItems,
       scope: 'dashboards'
@@ -166,7 +205,22 @@ const MainLayout = (props) => {
   const { children } = props;
   const { state, onBreakpoint, onCollapse } = useSider();
 
+  useCurrentUser();
+  const [location] = useLocation();
+  const { authToken } = useAuthToken();
+
+  useEffect(() =>{
+    if (!authToken) {
+      window.location.href = '/login';
+    }
+  }, [authToken, location.pathname]);
+
   const { restrictScopes } = usePermissions({});
+
+  const { currentUserState: currentUser } = useCurrentUserState();
+  const members = currentUser?.users_by_pk?.members || [];
+
+  const anyTeam = !!members.length;
 
   return (
     <Layout>
@@ -182,9 +236,15 @@ const MainLayout = (props) => {
           mode="vertical"
           restrictScopes={restrictScopes}
         />
+        <TeamMenu
+          mode="vertical"
+          anyTeam={anyTeam}
+          members={members}
+        />
         <UserMenu
           mode="vertical"
           restrictScopes={restrictScopes}
+          anyTeam={anyTeam}
         />
       </Sider>
       <Layout style={{ minHeight: 'auto' }}>
@@ -198,7 +258,10 @@ const MainLayout = (props) => {
             <MainMenu restrictScopes={restrictScopes} />
           </div>
           <div style={{ float: 'right' }}>
-            <UserMenu restrictScopes={restrictScopes} />
+            <UserMenu anyTeam={anyTeam} restrictScopes={restrictScopes} />
+          </div>
+          <div style={{ float: 'right' }}>
+            <TeamMenu anyTeam={anyTeam} members={members} />
           </div>
         </Header>
         <Content style={{ minHeight: 'auto', zIndex: 1 }}>
