@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 import { getOr } from 'unchanged';
 
-import { Link } from 'wouter';
+import { Link } from 'react-router-dom';
 import { Menu } from 'antd';
 
 import pickKeys from 'utils/pickKeys';
@@ -14,8 +14,8 @@ const MenuView = (props) => {
 
   const getLink = (node) => {
     const linkProps = {
-      href: node.path || '',
-      onClick: (node.onClick && (() => node.onClick(node))) || noop,
+      to: node.path || '',
+      onClick: (node.onClick && ((e) => node.onClick(node, e))) || noop,
     };
 
     return (
