@@ -1,6 +1,5 @@
 import { useEffect } from 'react';
-import useLocation from 'wouter/use-location';
-import { useTranslation } from 'react-i18next';
+import useLocation from 'hooks/useLocation';
 
 import { message } from 'antd';
 
@@ -26,7 +25,7 @@ const useCheckResponse = (response, cb = noop, meta = {}) => {
       response.data = null;
     }
   },
-    [response.data]
+  [cb, response.data, successMessage]
   );
 
   useEffect(() => {
@@ -43,7 +42,7 @@ const useCheckResponse = (response, cb = noop, meta = {}) => {
       response.error = null;
     }
   },
-    [response.error]
+  [cb, errorMessage, response.error, setLocation]
   );
 };
 
