@@ -1,167 +1,99 @@
-import { useCallback, useMemo, useEffect } from 'react';
+import { useEffect, useMemo } from 'react';
 
-import nanoid from 'nanoid';
-import { useMutation, useQuery } from 'urql';
-import { get, getOr } from 'unchanged';
-
-import trackEvent from 'utils/trackEvent';
-
-import useAuthContext from './useAuthContext';
+import useQuery from './useQuery';
+import useMutation from './useMutation';
 
 const newPinnedItemMutation = `
-  mutation NewPinnedItemMutation($input: CreatePinnedItemInput!) {
-    createPinnedItem(input: $input) {
-      pinnedItem {
-          id
-          name
-      }
+  mutation ($object: pinned_items_insert_input!) {
+    insert_pinned_items_one(object: $object) {
+      id
+      name
     }
   }
 `;
 
-const currentPinnedItemQuery = `
-  query CurrentPinnedItem($rowId: Int!) {
-    pinnedItemByRowId(rowId: $rowId) {
+const editPinnedItemMutation = `
+  mutation (
+    $pk_columns: pinned_items_pk_columns_input!,
+    $_set: pinned_items_set_input!
+  ) {
+    update_pinned_items_by_pk(pk_columns: $pk_columns, _set: $_set) {
       id
-      rowId
+    }
+  }
+`;
+
+const delPinnedItemMutation = `
+  mutation ($id: uuid!) {
+    delete_pinned_items_by_pk(id: $id) {
+      id
+    }
+  }
+`;
+
+const editPinnedItemQuery = `
+  query ($id: uuid!) {
+    pinned_items_by_pk(id: $id) {
+      id
+      created_at
+      updated_at
       name
       spec
-      specConfig
-      explorationByExplorationId {
-        datasourceId
-        slug
-        dataCube {
-          data
-          progress {
-            loading
-            stage
-            timeElapsed
-            error
-          }
-        }
-      }
-      dashboardByDashboardId {
-        rowId
-        name
-        layout
-      }
-    }
-  }
-`;
-
-const pinnedItemsQuery = `
-  query allPinnedItems {
-    allPinnedItems {
-      nodes {
+      spec_config
+      dashboard {
         id
-        rowId
+        layout
         name
       }
-    }
-  }
-`;
-
-const deletePinnedItemMutation = `
-  mutation DeletePinnedItemMutation($input: DeletePinnedItemByRowIdInput!) {
-    deletePinnedItemByRowId(input: $input) {
-      deletedPinnedItemId
-    }
-  }
-`;
-
-const updatePinnedItemMutation = `
-  mutation updatePinnedItem($input: UpdatePinnedItemInput!) {
-    updatePinnedItem(input: $input) {
-      pinnedItem {
-        name
+      exploration {
+        id
+        datasource_id
       }
     }
   }
 `;
 
-export default ({ rowId }) => {
-  const currentUser = useAuthContext();
+const role = 'user';
+export default (props = {}) => {
+  const { params = {} } = props;
+  const { editId } = params;
 
-  const [createMutation, executeNewMutation] = useMutation(newPinnedItemMutation);
-  const mExecuteNewMutation = useCallback(pinnedItem => {
-    const clientMutationId = nanoid();
-    trackEvent('Create pinned item');
+  const [createMutation, execCreateMutation] = useMutation(newPinnedItemMutation, { role });
+  const [updateMutation, execUpdateMutation] = useMutation(editPinnedItemMutation, { role });
+  const [deleteMutation, execDeleteMutation] = useMutation(delPinnedItemMutation, { role });
 
-    executeNewMutation({
-      input: {
-        clientMutationId,
-        pinnedItem: {
-          ...pinnedItem,
-          userId: currentUser.userId,
-        },
-      }
-    });
-  }, [currentUser.userId, executeNewMutation]);
-
-  const [allData, executeQueryAll] = useQuery({
-    query: pinnedItemsQuery,
-    pause: true,
-  });
-
-  const all = useMemo(() => getOr([], 'data.allPinnedItems.nodes', allData), [allData]);
-
-  const [currentData, executeQueryCurrent] = useQuery({
-    query: currentPinnedItemQuery,
+  const [currentData, execQueryCurrent] = useQuery({
+    query: editPinnedItemQuery,
     variables: {
-      rowId: parseInt(rowId, 10),
+      id: editId,
     },
     pause: true,
+  }, {
+    requestPolicy: 'cache-and-network',
+    role,
   });
 
-  const current = useMemo(() => get('data.pinnedItemByRowId', currentData) || {}, [currentData]);
-  const currentProgress = useMemo(
-    () => get('data.pinnedItemByRowId.explorationByExplorationId.dataCube.progress', currentData) || {},
-    [currentData]
-  );
-
-  const [deleteMutation, executeDeleteMutation] = useMutation(deletePinnedItemMutation);
-  const mExecuteDeleteMutation = useCallback(id => {
-    trackEvent('Delete Pinned Item');
-
-    executeDeleteMutation({
-      input: { rowId: id },
-    });
-  }, [executeDeleteMutation]);
-
-  const [updateMutation, execUpdateMutation] = useMutation(updatePinnedItemMutation);
-  const mExecUpdateMutation = useCallback((id, input = {}) => {
-    execUpdateMutation({
-      input: {
-        id,
-        pinnedItemPatch: input
-      },
-    });
-  }, [execUpdateMutation]);
+  const current = useMemo(() => currentData.data?.pinned_items_by_pk, [currentData.data]);
 
   useEffect(() => {
-    if (rowId || updateMutation.data) {
-      executeQueryCurrent({ requestPolicy: 'cache-and-network' });
+    if (editId) {
+      execQueryCurrent();
     }
-  }, [rowId, executeQueryCurrent, updateMutation.data]);
-
-  useEffect(() => {
-    if (currentProgress && currentProgress.loading) {
-      executeQueryCurrent({ requestPolicy: 'network-only' });
-    }
-  }, [currentProgress, executeQueryCurrent]);
+  }, [editId, execQueryCurrent]);
 
   return {
-    all,
     current,
-    currentProgress,
     queries: {
-      allData, executeQueryAll,
-      currentData
+      currentData,
+      execQueryCurrent,
     },
     mutations: {
-      createMutation, mExecuteNewMutation,
-      deleteMutation, mExecuteDeleteMutation,
-      updateMutation, mExecUpdateMutation
+      createMutation,
+      execCreateMutation,
+      deleteMutation,
+      execDeleteMutation,
+      updateMutation,
+      execUpdateMutation,
     },
   };
 };
