@@ -1,25 +1,24 @@
-import React, { useMemo, useEffect, useCallback, useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 import { useTranslation } from 'react-i18next';
-import { Empty, Icon, message, Popconfirm } from 'antd';
-
-import { getOr } from 'unchanged';
+import { Icon, Popconfirm, message } from 'antd';
 
 import GridLayout from 'react-grid-layout';
-import useLocation from 'wouter/use-location';
 
 import equals from 'utils/equals';
+
 import Loader from 'components/Loader';
 import ContentHeader from 'components/ContentHeader';
 import PinnedItem from 'components/PinnedItem';
 import EditableField from 'components/EditableField';
 import ErrorFound from 'components/ErrorFound';
 
+import useLocation from 'hooks/useLocation';
 import useCheckResponse from 'hooks/useCheckResponse';
 import useDashboards from 'hooks/useDashboards';
-import useAuth from 'hooks/useAuth';
 import usePermissions from 'hooks/usePermissions';
+import useAppSettings from 'hooks/useAppSettings';
 
 import 'react-grid-layout/css/styles.css';
 import 'react-resizable/css/styles.css';
@@ -27,15 +26,12 @@ import s from './Dashboards.module.css';
 
 export const rowHeight = 30;
 
-const Dashboards = ({ params }) => {
+const Dashboards = ({ match }) => {
   const { t } = useTranslation();
   const [, setLocation] = useLocation();
-  const basePath = '/d/dashboards';
+  const { withAuthPrefix } = useAppSettings();
+  const { params } = match;
   const { rowId } = params;
-
-  const {
-    setLastUsedDashboardId,
-  } = useAuth();
 
   const {
     all: dashboards,
@@ -44,39 +40,48 @@ const Dashboards = ({ params }) => {
     getItemGridData,
     queries: {
       allData,
-      executeQueryAll,
+      execQueryAll,
       currentData,
-      executeQueryCurrent
+      execQueryCurrent,
     },
     mutations: {
-      updateMutation, mExecUpdateMutation,
-      deleteMutation, mExecuteDeleteMutation
+      updateMutation,
+      execUpdateMutation,
+      deleteMutation,
+      execDeleteMutation,
     },
-  } = useDashboards({ editId: rowId, pauseQueryAll: false });
+  } = useDashboards({
+    params: {
+      editId: rowId,
+    },
+    pauseQueryAll: false,
+  });
 
-  useEffect(() => {
-    setLastUsedDashboardId(dashboard.rowId);
-  }, [dashboard?.rowId]);
-
-  const onDelete = async (res, err) => {
+  const onDelete = async (res) => {
     if (res) {
-      executeQueryAll();
+      execQueryAll();
 
       if (dashboards.length - 1 <= 0) {
         const firstDataSource = allData?.data?.allDatasources?.nodes?.[0]?.rowId;
-        setLastUsedDashboardId(null);
-        setLocation(`/d/explore/${firstDataSource}`);
+        setLocation(withAuthPrefix(`/explore/${firstDataSource}`));
       } else {
         const firstDashboard = allData?.data?.allDashboards?.nodes?.filter(
-          dashboard => dashboard.id !== res.deleteDashboard.deletedDashboardId
+          dash => dash.id !== res.deleteDashboard.deletedDashboardId
         )?.[0]?.rowId;
 
-        setLocation(`/d/dashboards/${firstDashboard}`);
+        setLocation(withAuthPrefix(`/dashboards/${firstDashboard}`));
       }
     }
   };
 
-  useCheckResponse(updateMutation, () => {}, {
+  const onUpdate = res => {
+    if (res) {
+      execQueryCurrent();
+      execQueryAll();
+    }
+  };
+
+  useCheckResponse(updateMutation, onUpdate, {
     successMessage: null,
   });
 
@@ -84,18 +89,47 @@ const Dashboards = ({ params }) => {
     successMessage: t('Successfully deleted'),
   });
 
-  const pinnedItems = useMemo(() => (
-    getOr([], 'pinnedItemsByDashboardId.nodes', dashboard)
-  ), [dashboard]);
-
   const onLayoutChange = (newLayout) => {
-    mExecUpdateMutation(dashboard.id, { name: dashboard.name, layout: newLayout });
+    let parsedLayout;
+
+    try {
+      parsedLayout = JSON.parse(JSON.stringify(newLayout));
+    } catch (err) {
+      message.error(err.toString());
+      return null;
+    }
+
+    if (!newLayout || equals(dashboard.layout, parsedLayout)) {
+      return null;
+    }
+
+    return execUpdateMutation({ 
+      pk_columns: {
+        id: dashboard.id,
+      },
+      _set: {
+        name: dashboard.name,
+        layout: newLayout,
+      },
+    });
   };
 
+
   const onRename = (id, value) => {
-    if (value.name !== dashboard.name) {
-      mExecUpdateMutation(id, value);
+    const { name } = value || {};
+
+    if (dashboard?.name && name === dashboard?.name) {
+      return null;
     }
+
+    return execUpdateMutation({ 
+      pk_columns: {
+        id,
+      },
+      _set: {
+        name,
+      },
+    });
   };
 
   const { fallback } = usePermissions({ scope: 'dashboards' });
@@ -103,7 +137,7 @@ const Dashboards = ({ params }) => {
     return fallback;
   }
 
-  if (!dashboard.rowId && !currentData.fetching) {
+  if (!dashboard.id && !currentData.fetching) {
     return <ErrorFound status={404} />;
   }
 
@@ -113,13 +147,13 @@ const Dashboards = ({ params }) => {
     <>
       <Loader spinning={currentData.fetching}>
         <ContentHeader
-          rowId={dashboard.rowId}
-          title={dashboard?.rowId && (
+          rowId={dashboard.id}
+          title={dashboard?.id && (
             <Loader spinning={nameLoading}>
               <div style={{ lineHeight: '32px', whiteSpace: 'nowrap' }}>
                 <EditableField
-                  id={dashboard.id}
-                  currentValue={dashboard.name}
+                  id={dashboard?.id}
+                  currentValue={dashboard?.name || ''}
                   renameFunc={onRename}
                   style={{ minWidth: '50px', fontWeight: '500' }}
                 />
@@ -129,11 +163,11 @@ const Dashboards = ({ params }) => {
           confirmTitle={t('Are you sure delete this dashboard?')}
           entities={dashboards}
           onChange={onChange}
-          extra={dashboard.rowId && (
+          extra={dashboard.id && (
             <div style={{ display: 'flex', flex: 1, justifyContent: 'flex-end' }}>
               <Popconfirm
                 title={t('Are you sure delete this dashboard?')}
-                onConfirm={() => mExecuteDeleteMutation(dashboard.id)}
+                onConfirm={() => execDeleteMutation({ id: dashboard.id })}
                 okText={t('Yes')}
                 okType="danger"
                 cancelText={t('No')}
@@ -154,11 +188,11 @@ const Dashboards = ({ params }) => {
           draggableCancel=".ant-typography, .marks, .vega-actions"
           width={1800}
         >
-          {pinnedItems.map((item) => (
-            <div className={s.item} key={item.rowId} data-grid={getItemGridData(item.rowId)}>
+          {dashboard?.pinned_items?.map((item) => (
+            <div className={s.item} key={item.id} data-grid={getItemGridData(item.id)}>
               <PinnedItem
-                rowId={item.rowId}
-                updateDashboard={executeQueryCurrent}
+                rowId={item.id}
+                updateDashboard={execQueryCurrent}
               />
             </div>
           ))}
@@ -169,11 +203,11 @@ const Dashboards = ({ params }) => {
 };
 
 Dashboards.propTypes = {
-  params: PropTypes.object,
+  match: PropTypes.object,
 };
 
 Dashboards.defaultProps = {
-  params: {},
+  match: {},
 };
 
 export default Dashboards;
