@@ -4,7 +4,8 @@ import PropTypes from 'prop-types';
 import { Button, Icon } from 'antd';
 
 import { useTranslation } from 'react-i18next';
-import useLocation from 'wouter/use-location';
+import useLocation from 'hooks/useLocation';
+import useAppSettings from 'hooks/useAppSettings';
 
 import DataSourceModal from 'components/DataSourceModal';
 import DataSourcesTable from 'components/DataSourcesTable';
@@ -15,10 +16,12 @@ import usePermissions from 'hooks/usePermissions';
 
 import iconDataSources from 'assets/images/icon_data_sources.svg';
 
-const DataSources = ({ params }) => {
+const DataSources = ({ match }) => {
   const { t } = useTranslation();
+  const { params = {} } = match || {};
   const [location, setLocation] = useLocation();
-  const basePath = '/d/sources';
+  const { withAuthPrefix } = useAppSettings();
+  const basePath = withAuthPrefix('/sources');
 
   const { fallback } = usePermissions({ scope: 'datasources' });
   if (fallback) {
@@ -26,7 +29,7 @@ const DataSources = ({ params }) => {
   }
 
   const onModalOpen = record => {
-    setLocation(`${basePath}/${record.rowId || ''}`);
+    setLocation(`${basePath}/${record.id || ''}`);
   };
 
   const onModalClose = () => {
@@ -37,12 +40,12 @@ const DataSources = ({ params }) => {
     setLocation(`${basePath}/new/${(dbType || '').toLowerCase()}`);
   };
 
-  const showNewForm = location.includes('/new');
+  const showNewForm = location.pathname.includes('/new');
 
   const breadcrumbs = [
     { path: basePath, title: 'Data Sources' },
     { path: `${basePath}/new`, title: 'New' },
-    params.dbType && { path: `${basePath}/new/${params.dbType}`, title: params.dbType },
+    params?.dbType && { path: `${basePath}/new/${params?.dbType}`, title: params?.dbType },
   ].filter(v => !!v);
 
   return (
@@ -68,7 +71,7 @@ const DataSources = ({ params }) => {
         title={t('New Data Source')}
         dataSource={{}}
         initialValues={{
-          dbType: (params.dbType || '').toUpperCase(),
+          db_type: (params?.dbType || '').toUpperCase(),
         }}
         breadcrumbs={breadcrumbs}
         onChange={onConnectNewClick}
@@ -77,7 +80,7 @@ const DataSources = ({ params }) => {
         visible={showNewForm}
       />
       <DataSourcesTable
-        editId={params.rowId}
+        editId={params?.rowId}
         onModalOpen={onModalOpen}
         onModalClose={onModalClose}
       />
@@ -86,7 +89,7 @@ const DataSources = ({ params }) => {
 };
 
 DataSources.propTypes = {
-  params: PropTypes.object.isRequired,
+  match: PropTypes.object.isRequired,
 };
 
 DataSources.defaultProps = {};
