@@ -1,69 +1,90 @@
 import React, { useCallback, useEffect } from 'react';
 import PropTypes from 'prop-types';
+import { useTrackedEffect } from 'ahooks';
 
-import useLocation from 'wouter/use-location';
 import { useTranslation } from 'react-i18next';
 
-import { getOr } from 'unchanged';
-
-import { Link } from 'wouter';
+import { Link } from 'react-router-dom';
 import { Empty, Button } from 'antd';
 
+import equals from 'utils/equals';
 import Loader from 'components/Loader';
 import ContentHeader from 'components/ContentHeader';
 import ExploreWorkspace from 'components/ExploreWorkspace';
+import ErrorFound from 'components/ErrorFound';
 
-import useDataSources from 'hooks/useDataSources';
-import useAuth from 'hooks/useAuth';
-import useDataSchemasSubscription from 'hooks/useDataSchemasSubscription';
+import useLocation from 'hooks/useLocation';
+import useAppSettings from 'hooks/useAppSettings';
+import useCurrentUserState from 'hooks/useCurrentUserState';
+import useSources from 'hooks/useSources';
 import usePermissions from 'hooks/usePermissions';
 
 const Explore = (props) => {
   const { t } = useTranslation();
   const [, setLocation] = useLocation();
+  const { withAuthPrefix } = useAppSettings();
+  const { match } = props;
+  const { params = {} } = match;
 
   const urlParams = new URLSearchParams(window.location.search);
   const tabId = urlParams.get('tabId');
   const chartId = urlParams.get('chart');
 
-  const [dataSourceId, explorationId, taskId] = (props?.params?.rest || '').split('/');
-  const basePath = ['/d/explore', dataSourceId, explorationId].filter(v => !!v).join('/');
+  const [dataSourceId, explorationId] = (params?.rest || '').split('/');
+  const basePath = [withAuthPrefix('/explore'), dataSourceId, explorationId].filter(v => !!v).join('/');
+
+  const { currentUserState: currentUser } = useCurrentUserState();
 
   const {
-    lastUsedDataSourceId,
-    setLastUsedDataSourceId
-  } = useAuth();
-
-  const {
-    all: dataSources,
-    current: dataSource,
+    current,
+    currentMeta,
     queries: {
-      currentData: {
-        fetching: loadingDataSource
-      },
-      executeQueryCurrent: loadDataSource,
+      metaData,
+      currentData,
+      execQueryMeta,
     },
-  } = useDataSources({ editId: dataSourceId });
-
-  useDataSchemasSubscription(() => {
-    if (dataSourceId) {
-      loadDataSource({ requestPolicy: 'network-only' });
-    }
+  } = useSources({
+    params: {
+      editId: dataSourceId,
+    },
+    pauseQueryAll: true,
   });
 
-  const onChange = useCallback((key = dataSource.rowId) => {
-    setLocation(`/d/explore/${key}`);
-  }, [dataSource.rowId, setLocation]);
+  const fetching = currentData.fetching || metaData.fetching;
+
+  useTrackedEffect((changes, previousDeps, currentDeps) => {
+    const prevData = previousDeps?.[0];
+    const currData = currentDeps?.[0];
+
+    let dataDiff = false;
+    if (!prevData || !currData) {
+      dataDiff = false;
+    } else {
+      dataDiff = !equals(prevData, currData);
+    }
+
+    if (dataDiff) {
+      execQueryMeta({ requestPolicy: 'network-only' });
+    }
+  }, [currentUser.dataschemas, execQueryMeta]);
 
   useEffect(() => {
-    if (dataSourceId && lastUsedDataSourceId !== dataSourceId) {
-      setLastUsedDataSourceId(dataSourceId);
+    if (dataSourceId) {
+      execQueryMeta({ requestPolicy: 'network-only' });
     }
-  }, [dataSourceId, lastUsedDataSourceId, setLastUsedDataSourceId]);
+  }, [dataSourceId, execQueryMeta]);
+
+  const onChange = useCallback((key = current.id) => {
+    setLocation(withAuthPrefix(`/explore/${key}`));
+  }, [current.id, setLocation, withAuthPrefix]);
 
   const { fallback } = usePermissions({ scope: 'explore' });
   if (fallback) {
     return fallback;
+  }
+
+  if (currentData?.data?.datasources_by_pk === null) {
+    return <ErrorFound status={404} />;
   }
 
   if (!dataSourceId) {
@@ -89,23 +110,22 @@ const Explore = (props) => {
   }
 
   return (
-    <Loader spinning={loadingDataSource}>
+    <Loader spinning={fetching}>
       <ExploreWorkspace
         basePath={basePath}
         header={(
           <ContentHeader
-            rowId={dataSource.rowId}
-            title={dataSource.name || 'Select DataSource'}
-            entities={dataSources}
+            selectedId={current.id}
+            title={current.name || 'Select DataSource'}
+            entities={currentUser?.datasources}
             onChange={onChange}
           />
         )}
-        dataSource={dataSource}
-        loading={loadingDataSource}
+        source={current}
+        meta={currentMeta}
         params={({
           dataSourceId,
           explorationId,
-          taskId,
           tabId,
           chartId,
         })}
@@ -115,11 +135,11 @@ const Explore = (props) => {
 };
 
 Explore.propTypes = {
-  params: PropTypes.object,
+  match: PropTypes.object,
 };
 
 Explore.defaultProps = {
-  params: {},
+  match: {},
 };
 
 export default Explore;
