@@ -4,11 +4,11 @@ import PageInfo from 'components/PageInfo';
 import Container from 'components/Container';
 import UserSettings from 'components/UserSettings';
 
-import useAuthContext from 'hooks/useAuthContext';
+import useCurrentUserState from 'hooks/useCurrentUserState';
 
 const Profile = () => {
-  const currentUser = useAuthContext();
-  const title = `${currentUser.name} (${currentUser.email})`;
+  const { currentUserState } = useCurrentUserState();
+  const title = `${currentUserState.users_by_pk?.full_name || ''} (${currentUserState.users_by_pk?.display_name})`.trim();
 
   return (
     <Container>
